# packages/shared/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ghar Khoj - Broker-Free Room Rental Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rooms:
    get:
      operationId: getRooms
      tags: [rooms]
      summary: Get all rooms with optional filters
      parameters:
        - name: city
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: roomType
          in: query
          schema:
            type: string
        - name: tenantType
          in: query
          schema:
            type: string
        - name: parking
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of rooms
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomListResponse"
    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a new room listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoomInput"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/owner/{ownerId}:
    get:
      operationId: getRoomsByOwner
      tags: [rooms]
      summary: Get rooms belonging to a specific owner
      parameters:
        - name: ownerId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Owner's rooms
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomListResponse"
  /rooms/{id}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get a single room by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Room details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Room not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateRoom
      tags: [rooms]
      summary: Update a room listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoomInput"
      responses:
        "200":
          description: Updated room
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
    delete:
      operationId: deleteRoom
      tags: [rooms]
      summary: Delete a room listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Room deleted
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users:
    post:
      operationId: createUser
      tags: [users]
      summary: Create / register a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlBody"
      responses:
        "200":
          description: Presigned upload URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrlResponse"
  /tenant-preferences/{userId}:
    get:
      operationId: getTenantPreferences
      tags: [matches]
      summary: Get tenant preferences
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tenant preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantPreference"
    put:
      operationId: upsertTenantPreferences
      tags: [matches]
      summary: Create or update tenant preferences
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertTenantPreferenceInput"
      responses:
        "200":
          description: Preferences saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantPreference"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error, message]
    Room:
      type: object
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        title:
          type: string
        description:
          type: string
        price:
          type: number
        roomType:
          type: string
          enum: [single, double, flat, studio, shared]
        tenantType:
          type: string
          enum: [any, family, student, male, female, professional]
        city:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        parking:
          type: boolean
        amenities:
          type: array
          items:
            type: string
        photos:
          type: array
          items:
            type: string
        isVerified:
          type: boolean
        isAvailable:
          type: boolean
        nearbyLandmarks:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - ownerId
        - title
        - price
        - roomType
        - tenantType
        - city
        - address
        - latitude
        - longitude
        - parking
        - amenities
        - photos
        - isVerified
        - isAvailable
        - createdAt
    CreateRoomInput:
      type: object
      properties:
        ownerId:
          type: integer
        title:
          type: string
        description:
          type: string
        price:
          type: number
        roomType:
          type: string
          enum: [single, double, flat, studio, shared]
        tenantType:
          type: string
          enum: [any, family, student, male, female, professional]
        city:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        parking:
          type: boolean
        amenities:
          type: array
          items:
            type: string
        photos:
          type: array
          items:
            type: string
        nearbyLandmarks:
          type: array
          items:
            type: string
      required:
        - ownerId
        - title
        - price
        - roomType
        - tenantType
        - city
        - address
        - latitude
        - longitude
        - parking
    RoomListResponse:
      type: object
      properties:
        rooms:
          type: array
          items:
            $ref: "#/components/schemas/Room"
        total:
          type: integer
      required: [rooms, total]
    User:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [tenant, owner, admin]
        isVerified:
          type: boolean
        verificationStatus:
          type: string
          enum: [pending, approved, rejected, none]
        profilePhoto:
          type: string
        bio:
          type: string
        preferredCity:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - firstName
        - email
        - role
        - isVerified
        - verificationStatus
        - createdAt
    CreateUserInput:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [tenant, owner, admin]
        preferredCity:
          type: string
      required: [firstName, email, role]
    UpdateUserInput:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        bio:
          type: string
        preferredCity:
          type: string
        profilePhoto:
          type: string
...
    UpdateRoomInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        roomType:
          type: string
        tenantType:
          type: string
        parking:
          type: boolean
        amenities:
          type: array
          items:
            type: string
        photos:
          type: array
          items:
            type: string
        isAvailable:
          type: boolean
        nearbyLandmarks:
          type: array
          items:
            type: string
...
    RequestUploadUrlBody:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
      required: [name, contentType]
    RequestUploadUrlResponse:
      type: object
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          type: object
          properties:
            name:
              type: string
            size:
              type: integer
            contentType:
              type: string
      required: [uploadURL, objectPath]
    TenantPreference:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        roomType:
          type: string
        tenantType:
          type: string
        city:
          type: string
        minBudget:
          type: number
        maxBudget:
          type: number
        parking:
          type: boolean
        amenities:
          type: array
          items:
            type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, userId, createdAt, updatedAt]
    UpsertTenantPreferenceInput:
      type: object
      properties:
        roomType:
          type: string
        tenantType:
          type: string
        city:
          type: string
        minBudget:
          type: number
        maxBudget:
          type: number
        parking:
          type: boolean
        amenities:
          type: array
          items:
            type: string
        notes:
          type: string
...